clipboard-apis/setdata_does_not_modify_text_about_to_be_inserted_in_a_paste_event_w_preventdefault_paste_on_document.html

<!DOCTYPE html>
<html><head>
    <meta charset="UTF-8">
    <title>Clipboard API - setData() does not modify text about to be inserted in a paste event - w preventDefault()  - paste on document</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="support/_lib.js"></script>
    <script>
/** setData() does not modify text about to be inserted in a paste event - w preventDefault()  */
/* Events: paste */
/* paste data: "pasted text" */
/* Test HTML: <form><input autofocus></form> */
function clipboard_api_test(e, test_obj){
    test_obj.step_timeout(function(){
      result(function(){
        assert_equals(e.target.value, 'pasted text');
      });
    }, 10);
    e.clipboardData.setData('text/plain', 'modified text');
    e.preventDefault();
}

    </script>
</head>
<body>

    <p>FAILED (This TC requires JavaScript enabled)</p>
    <form><input autofocus></form>
    <div id="log"></div>
    <script>
    (function(){
        var t = async_test()
        var dataToPaste="pasted text" ;
        var extPassCond=null;
        var eventTarget=document;
        var eventType='paste';
        setup(function(){
                setupTest(t, eventTarget, eventType, dataToPaste, extPassCond );
            });
    })();
    </script>
</body></html>
